requirements.txt
google-genai
python-dotenv
pytest
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pgvector